Basic perf comparison against uap-cpp
Although uap-cpp apparently uses a bespoke prefiltering scheme rather
than `FilteredRE2`, even though it already depends on re2. The idea
seems conceptually similar, but it might be slower? It seems to be
using some sort of trie versus a `RE2::Set` or an aho-corasick impl,
so might be slower because of that.

Anyway that means we'll probably still need to bench against
`FilteredRE2` somehow to actually get 1:1 data... ugh...