Update uap-cpp comparison following regex-filtered optimisations
Also fix the repeat counts, weirdly enough 1000 iterations is not
10x *faster* than 100 iterations.

And in confirming that I had had my repeat counts mixed up I got very
slightly better runtimes for the uap-cpp bench, so put that in even if
the difference is minute.

Also add the machine I'm using for reference.