Implement bespoke cut down IntMap
As noted in 4f1c7df94f4171d78e0128146254d99e6a12c27e, turns out it's
trivial enough I could well have done that in the first place after
validating that was the issue, rather than add nohash and keep
indexmap.

This is a highly cut down implementation to just the operations
necessary for `atom_to_re` and `propagate_match`. The gain in runtime
are near 10% which is nothing to sneeze at. Though with almost 20%
less instructions (!!!) that means the IPC has taken a dive and is now
a hair below 3. I still have essentially no idea what that means, but
I think it bears noting anyway.

```
       42.63 real        42.49 user         0.02 sys
           145833984  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
                9015  page reclaims
                   6  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                   0  messages sent
                   0  messages received
                   0  signals received
                   0  voluntary context switches
                 118  involuntary context switches
        408805706973  instructions retired
        137020670968  cycles elapsed
           140592576  peak memory footprint
```