Optimise pyo3 use
While this has not been profiled, it seems rather unlikely that the
python implementation uses the same allocator as Rust (whether
system or other). This means *returning* a `String` is an extra
allocation as that will need to be copied again to a `PyString` on the
ouput. Which is just dumb.

Copy the internal `Cow` directly to a `PyString` on output. On input,
a `String` is necessary as there's no way to lifetime properly from
Python, so the extractors need to be `'static`.