In light of the previous commit, revert atom length for UAs and OS
Keep devices at 2 as it benefits *massively*, but UAs and OS seem to
either not benefit or be slightly hampered by smaller atoms.

Would be a good idea to investigate why eventually, and maybe make the
bench scripts more flexible so they can be more easily run across the
three categories and with varied atom lengths.